\name{miRBase_20}
\alias{miRBase_20}
\docType{data}
\title{ miRBase version 20 mature sequences (homo sapiens) }
\description{
Mature miRNA sequences of Homo Sapiens from miRBase version .
}
\usage{data(miRBase_20)}
\source{
The microRNA Registry.
Griffiths-Jones S. 
NAR 2004 32(Database Issue):D109-D111
}
\keyword{datasets}