\name{OPIrsa}
\alias{OPIrsa}
\title{
OPIrsa
}
\description{
look Konig paper/code for explanation about this method
}
\usage{
OPIrsa(Groups,Scores,opts,Data=NULL)
}
\arguments{
  \item{Groups}{
    look Konig paper/code for explanation about this parameter
}
  \item{Scores}{
    look Konig paper/code for explanation about this parameter
}
  \item{opts}{
    look Konig paper/code for explanation about this parameter
}
  \item{Data}{
    look Konig paper/code for explanation about this parameter
}
}
\value{
look Konig paper/code for explanation about this return value
}
\references{
A probability-based approach for the analysis of large-scale RNAi screens. 
Renate Konig et al. Nature Methods 2007
}
\author{
Andrea Franceschini
}
