\name{OPIrsaScore}
\alias{OPIrsaScore}
\title{
OPIrsaScore
}
\description{
look Konig paper in order to have information about this function
}
\usage{
OPIrsaScore(I_rank, N, i_min=1, i_max=-1)
}
\arguments{
  \item{I_rank}{
    look Konig paper in order to have information about this parameter
}
  \item{N}{
    look Konig paper in order to have information about this parameter
}
  \item{i_min}{
    look Konig paper in order to have information about this parameter
}
  \item{i_max}{
    look Konig paper in order to have information about this parameter
}
}
\value{
look Konig paper in order to have information about this return value
}
\references{
A probability-based approach for the analysis of large-scale RNAi screens. 
Renate Konig et al. Nature Methods 2007
}
\author{
Andrea Franceschini
}
