\name{uuk_screen_dh}
\alias{uuk_screen_dh}
\docType{data}
\title{cell number phenotype of the Uukuniemi DHARMACON siRNA genome wide screen }
\description{
  Cell Number phenotype of a genome wide siRNA screen that have been performed on Hela DC-SIGN cells that have been infected with the Uukuniemi virus.
  The screen has been performed using a DHARMACON pooled library with pools of 4 siRNA oligos x gene.
}
\usage{data(uuk_screen_dh)}
\format{
  Data frame with 70304 observations on the following 3 variables.
  \describe{
    \item{\code{GeneID}}{a numeric vector}
    \item{\code{siRNA_seq}}{a character vector}
    \item{\code{score}}{a numeric vector}
  }
}
\source{
Roger Meier et al. 2014
}
\keyword{datasets}